Refactor: extract DataFramer class to module
